:art: Use specific uint types

+++ include/emit.h
@@ -24,9 +24,9 @@ struct EmitterInfo_ {
   char *escape;
   VM_Code (*emit_code)(const Emitter);
   VM_Code code;
-  uint memoize;
-  uint unroll;
-  uint line;
+  uint16_t memoize;
+  uint16_t unroll;
+  uint16_t line;
   bool debug;
 };
